Update nf-shellapi-shellexecutew.md (#1980)
* Update nf-shellapi-shellexecutew.md

Undocumented behaviour in ShellExecuteW:

If lpFile specifies an executable file and "open" or "runas" verbs are used in lpOperation, a whitespace character is appended to the commandline of the new created process.
If lpParameters is not null, the whitespace is added between the end of the executable name and lpParameters string in the commandline of the new created process.

Suggestion:

Document that ShellExecuteW appends a whitespace character to the commandline of the created process if lpFile specifies an executable file and open or runas verbs are used for lpOperation.

* Update nf-shellapi-shellexecutew.md

Document that the command line string of a process created with "open" or "runas" verbs specified in lpOperation includes a trailing whitespace character and if parameters are passed to the application with lpParameters, they are appended to that command line.

* Update nf-shellapi-shellexecutew.md

Better wording.