Merged PR 21750: Add API doc for EnclaveUsesAttestedKeys in winenclaveapi.h
Add API doc for EnclaveUsesAttestedKeys in winenclaveapi.h.

Let me know if we want to provide all this information, or if anything should be held back.

Any See Also links for related APIs we would add to the doc?

----
#### AI description  (iteration 1)
#### PR Classification
Documentation

#### PR Summary
This pull request adds documentation for the `EnclaveUsesAttestedKeys` function in the `winenclaveapi.h` header file.
- Added a new file `/sdk-api-src/content/winenclaveapi/nf-winenclaveapi-enclaveusesattestedkeys.md` documenting the `EnclaveUsesAttestedKeys` function, including its description, return type, and remarks on its functionality.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->